writable3

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -238,9 +238,9 @@ public class PairsPMI  extends Configured implements Tool {
     }
   }
 
-  private static final class MyPartitioner extends Partitioner<PairOfStrings, IntWritable> {
+  private static final class MyPartitioner extends Partitioner<PairOfStrings, FloatWritable> {
     @Override
-    public int getPartition(PairOfStrings key, IntWritable value, int numReduceTasks) {
+    public int getPartition(PairOfStrings key, FloatWritable value, int numReduceTasks) {
       return (key.getLeftElement().hashCode() & Integer.MAX_VALUE) % numReduceTasks;
     }
   }
